// client/src/app/models/visita.ts
export interface Visita {
  id?: string;
  idProfesional?: string;
  idpaciente?: string;
  paciente?: string;
  temperatura?: string;
  peso?: string;
  presionArterial?: string;
  fecha?: string;
  hora?: string;
  medicamento?: string;
  idMedicamento?: string;
  laboratorio?: string;
  idlaboratorio?: string;
  dosisDiaria?: string;
  observaciones?: string;
}